Move timeout test to the SDK
The timeout test doesn't belong to the core, it belongs to the SDK.
This PR moves the test to the appropriate section in SDK.
